+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Fix `resize` implementation of resiable queue
+
 ## [0.2.1] - 2022-03-11
 
 ### Added
@@ -24,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - First release
 
+[unreleased]: https://github.com/bikeshedder/deadqueue/compare/0.2.1...HEAD
 [0.1.0]: https://github.com/bikeshedder/deadqueue/releases/tag/v0.1.0
 [0.2.0]: https://github.com/bikeshedder/deadqueue/releases/tag/v0.2.0
 [0.2.1]: https://github.com/bikeshedder/deadqueue/releases/tag/v0.2.1